rename the project
