+++ examples/multiple-alternative-audio-tracks/index.html
@@ -16,7 +16,7 @@
       </select>
     </div>
     <script src="/node_modules/video.js/dist/video.js"></script>
-    <script src="/dist/videojs-contrib-hls.js"></script>
+    <script src="/dist/videojs-http-streaming.js"></script>
     <script>
       (function(window, videojs) {
         var player = window.player = videojs('maat-player');
