+++ index.html
@@ -2,7 +2,7 @@
 <html>
 <head>
   <meta charset="utf-8">
-  <title>videojs-contrib-hls Demo</title>
+  <title>videojs-http-streaming Demo</title>
   <link href="/node_modules/video.js/dist/video-js.css" rel="stylesheet">
   <style>
     body {
@@ -29,7 +29,7 @@
     <p>The video below is an <a href="https://developer.apple.com/library/ios/documentation/networkinginternet/conceptual/streamingmediaguide/Introduction/Introduction.html#//apple_ref/doc/uid/TP40008332-CH1-SW1">HTTP Live Stream</a>. On desktop browsers other than Safari, the HLS plugin will polyfill support for the format on top of the video.js Flash tech.</p>
     <p>Due to security restrictions in Flash, you will have to load this page over HTTP(S) to see the example in action.</p>
   </div>
-  <video id="videojs-contrib-hls-player" class="video-js vjs-default-skin" controls>
+  <video id="videojs-http-streaming-player" class="video-js vjs-default-skin" controls>
     <source src="https://d2zihajmogu5jn.cloudfront.net/bipbop-advanced/bipbop_16x9_variant.m3u8" type="application/x-mpegURL">
   </video>
 
@@ -47,10 +47,10 @@
   </ul>
 
   <script src="/node_modules/video.js/dist/video.js"></script>
-  <script src="/dist/videojs-contrib-hls.js"></script>
+  <script src="/dist/videojs-http-streaming.js"></script>
   <script>
     (function(window, videojs) {
-      var player = window.player = videojs('videojs-contrib-hls-player');
+      var player = window.player = videojs('videojs-http-streaming-player');
 
       // hook up the video switcher
       var loadUrl = document.getElementById('load-url');
