+++ scripts/watch-test.js
@@ -13,11 +13,11 @@ glob('test/**/*.test.js', function(err, files) {
   .transform('browserify-shim', {global: true});
 
   var bundle = function() {
-    b.bundle().pipe(fs.createWriteStream('dist-test/videojs-contrib-hls.js'));
+    b.bundle().pipe(fs.createWriteStream('dist-test/videojs-http-streaming.js'));
   };
 
   b.on('log', function(msg) {
-    process.stdout.write(msg + ' dist-test/videojs-contrib-hls.js\n');
+    process.stdout.write(msg + ' dist-test/videojs-http-streaming.js\n');
   });
 
   b.on('update', bundle);
