+++ test/browserify-test.js
@@ -1,6 +1,6 @@
 /* eslint-disable no-var */
 /* eslint-env qunit */
-var hls = require('../es5/videojs-contrib-hls.js');
+var hls = require('../es5/videojs-http-streaming.js');
 var q = window.QUnit;
 
 q.module('Browserify Require');
