+++ test/index.html
@@ -16,7 +16,7 @@
   <script src="/dist-test/webpack-test.js"></script>
   <script src="/node_modules/video.js/dist/video.js"></script>
   <script src="/node_modules/videojs-flash/dist/videojs-flash.js"></script>
-  <script src="/dist-test/videojs-contrib-hls.js"></script>
+  <script src="/dist-test/videojs-http-streaming.js"></script>
 
 </body>
 </html>
