+++ test/karma.conf.js
@@ -38,7 +38,7 @@ module.exports = function(config) {
       'node_modules/videojs-flash/dist/videojs-flash.js',
       'dist-test/browserify-test.js',
       'dist-test/webpack-test.js',
-      'dist-test/videojs-contrib-hls.js'
+      'dist-test/videojs-http-streaming.js'
     ],
     browserConsoleLogOptions: {
       level: 'error',
