+++ utils/switcher/index.html
@@ -184,7 +184,7 @@
 
   <div class="footer-container">
     <footer class="wrapper">
-      <h3>videojs-contrib-hls</h3>
+      <h3>videojs-http-streaming</h3>
     </footer>
   </div>
   <div id="qunit-fixture"></div>
